+++ contrib/opencensus-ext-azure/CHANGELOG.md
@@ -12,6 +12,8 @@
   ([#946](https://github.com/census-instrumentation/opencensus-python/pull/946))
 - Added queue capacity configuration for exporters
   ([#949](https://github.com/census-instrumentation/opencensus-python/pull/949))
+- Disable heartbeat metrics in exporters
+  ([#949](https://github.com/census-instrumentation/opencensus-python/pull/949))
 
 ## 1.0.4
 Released 2020-06-29