+++ contrib/opencensus-ext-azure/CHANGELOG.md
@@ -2,6 +2,14 @@
 
 ## Unreleased
 
+## 1.0.6
+Released 2021-01-14
+
+- Disable heartbeat metrics in exporters
+  ([#984](https://github.com/census-instrumentation/opencensus-python/pull/984))
+- Loosen instrumentation key validation to GUID
+  ([#984](https://github.com/census-instrumentation/opencensus-python/pull/984))
+
 ## 1.0.5
 Released 2020-10-13